added text-align to css and moved constants.jsx into /src/ '
