+++ aunt-leahs-app/src/App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import logo from './logo.svg';
 import './App.css';
 import Head from './components/header/header.jsx';
-import constants from './components/constants.jsx';
+import constants from './constants.jsx';
 
 const subText = constants.headerConstants.subText;
 function App() {
