+++ aunt-leahs-app/src/styles.css
@@ -4,4 +4,5 @@
 }
 .header{
     color:black;
+    text-align: center;
 }
